refactor(previewpdf): refactored preview pdf functionality to a function in utils.ts
